requirements.txt
bcrypt==3.2.2
blinker==1.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
requests==2.28.1
SQLAlchemy==1.4.37
typing-extensions==4.2.0
urllib3==1.26.10
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
